test: enable FindInFiles-integ-test
